Gateway can be omitted from pool spec
This change allows the Gateway to be omitted from the Pool, preventing
an error. Also, prior to this change, Gateway was required when specifying
`pool.spec.addresses`.

We took this opportunity to move what we considered business logic out
of the pool utility package, and brought it back closer to the other
business logic in the controller. The gateway is conditionally added to
the list of in use IPs in the controller.

fixes: #89

Signed-off-by: Tyler Schultz <tschultz@vmware.com>
Co-authored-by: Aidan Obley <aobley@vmware.com>
Co-authored-by: Edwin Xie <exie@vmware.com>