Fix GlobalPools determination of in use IPs
The namespace of a Global pool is empty string. This commit changes the
namespace parameter of the lookup for in use IPs to use the pool
namespace instead of the claim namespace.

In the case of InClusterIPPools the pool namespace matches the claim
namespace. This namespace paramter is supplied to the in use IP lookup,
and remains effectively the same.

In the case of GlobalInClusterIPPools, using the pool namespace will
result in an empty namespace string. This empty namespace string will be
used to find in use IPs. The empty string causes the lookup to find
addresses in all namespaces, which is needed in the case of a global
pool.

This commit addresses: https://github.com/telekom/cluster-api-ipam-provider-in-cluster/issues/98

Signed-off-by: Tyler Schultz <tschultz@vmware.com>
Co-authored-by: Edwin Xie <exie@vmware.com>